+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ v3.1.0 (UNRELEASED)
 
 - Improve handling of paths with arbitrary encodings in the scanner. (#20, PR: #29)
 - Add ``.cue`` to the list of excluded file extensions. (PR: #29)
+- Replace ``os.path`` usage with ``pathlib`` to handle arbitrary file path
+  encodings better. (#20, PR: #30)
 
 
 v3.0.0 (2019-12-22)